requirements.txt
asgiref==3.3.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.7
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-import-export==2.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
docx==0.2.4
docxtpl==0.11.5
drf-renderer-xlsx==0.4.0
et-xmlfile==1.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupPy==1.14
MarkupSafe==2.0.1
numpy==1.20.3
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.3.0
psycopg2-binary==2.8.6
PyJWT==2.0.1
python-docx==0.8.11
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
XlsxWriter==1.4.3
xlwt==1.3.0
